Added auto emission and transmission colors in the Material. Added SamplesLight CustomNumberPicker. Changed AreaLight to a triangle shape. Added the -g3 flag to the cmake files. Added light blue front wall in the cornellBoxScene2. Added a sphere that refracts light in the cornellBoxScene2. Added a green triangle in the cornellBoxScene2. Added 2 more spheres blue and yellow in the spheresScene2. Added a method in the Ray to calculate the symmetric direction.

+++ jniLibs/Components/src/main/cpp/MobileRT/Samplers/Stratified.h
@@ -10,7 +10,7 @@
 namespace MobileRT {
     class Stratified : public Sampler {
     public:
-        explicit Stratified(const unsigned int domainSize, const unsigned int samples);
+        explicit Stratified(const unsigned long long int domainSize, const unsigned int samples);
 
         explicit Stratified(const unsigned int width, const unsigned int height,
                             const unsigned int samples,
